clickable background image and profile image Done

+++ API/Loader.swift
@@ -13,7 +13,7 @@ import SwiftyJSON
 // loader class from network ..........
 final class Loader {
     
-    class func getAllFollowersFromServer(cursor: Int, withURL url: String, parameter: [String: Any], token: HTTPHeaders, completion: @escaping (_ followers: [Follower]?, _ success: Bool?, _ error: Error?, _ next_cursor: Int) -> ()) {
+    class func getAllFollowersFromServer(cursor: Int64, withURL url: String, parameter: [String: Any], token: HTTPHeaders, completion: @escaping (_ followers: [Follower]?, _ success: Bool?, _ error: Error?, _ next_cursor: Int64) -> ()) {
         Alamofire.request(url, method: .get, parameters: parameter, encoding: URLEncoding.default, headers: token)
         .validate(statusCode: 200..<300)
             .responseJSON { (response) in
@@ -29,7 +29,7 @@ final class Loader {
                             followers.append(follower)
                         }
                     }
-                    let next_cursor = json["next_cursor"].int ?? cursor
+                    let next_cursor = json["next_cursor"].int64 ?? cursor
                     print("next_cursor is \(next_cursor)---------")
                     completion(followers,true, nil, next_cursor)
